# Optimize `BuildBlobSidecars` Merkle proof computation by pre-computing subtrees