--- a/changelog/tt_sushi.md
+++ b/changelog/tt_sushi.md
@@ -1,0 +1,2 @@
+# Changed
+Optimize proposer inclusion proof calcuation by pre caching subtries