--- a/changelog/tt_sushi.md
+++ b/changelog/tt_sushi.md
@@ -1,2 +1,3 @@
-# Changed
-Optimize proposer inclusion proof calcuation by pre caching subtries
+### Changed
+
+- Optimize proposer inclusion proof calcuation by pre caching subtries