This PR fixes #13254

## Summary
Optimizes `BuildBlobSidecars` by pre-computing shared Merkle tree components once instead of recalculating for each blob, eliminating O(n) redundant computations.

## Changes
  - Add `PrecomputeMerkleProofComponents()` to cache expensive subtree operations
  - Add `MerkleProofKZGCommitmentFromComponents()` for efficient index-specific proofs
  - Update `BuildBlobSidecars()` to use pre-computed components

  ## Performance Impact
  | Blob Count | Speed Improvement | Memory Reduction |
  |------------|-------------------|------------------|
  | 3 blobs    | **2.12x faster**  | 53% less memory  |
  | 9 blobs    | **5.16x faster**  | 77% less memory  |

```
goos: darwin
goarch: arm64
pkg: github.com/OffchainLabs/prysm/v6/beacon-chain/rpc/prysm/v1alpha1/validator
cpu: Apple M4 Pro
BenchmarkBuildBlobSidecars_Original_1Blob-14              106045             10217 ns/op           17900 B/op        238 allocs/op
BenchmarkBuildBlobSidecars_Optimized_1Blob-14             116745             10210 ns/op           17951 B/op        239 allocs/op
BenchmarkBuildBlobSidecars_Original_2Blobs-14              71059             16859 ns/op           31400 B/op        440 allocs/op
BenchmarkBuildBlobSidecars_Optimized_3Blobs-14            106286             11249 ns/op           21955 B/op        282 allocs/op
BenchmarkBuildBlobSidecars_Original_3Blobs-14              49640             24587 ns/op           46355 B/op        661 allocs/op
BenchmarkBuildBlobSidecars_Optimized_4Blobs-14            104604             11458 ns/op           23420 B/op        300 allocs/op
BenchmarkBuildBlobSidecars_Original_9Blobs-14              16446             72943 ns/op          151024 B/op       2206 allocs/op
BenchmarkBuildBlobSidecars_Optimized_9Blobs-14             81859             14643 ns/op           34711 B/op        423 allocs/op
PASS
ok      github.com/OffchainLabs/prysm/v6/beacon-chain/rpc/prysm/v1alpha1/validator      11.653s

goos: darwin
goarch: arm64
pkg: github.com/OffchainLabs/prysm/v6/beacon-chain/rpc/prysm/v1alpha1/validator
cpu: Apple M4 Pro
BenchmarkMerkleProofKZGCommitment_Original-14              43825             26859 ns/op           54335 B/op        824 allocs/op
BenchmarkMerkleProofKZGCommitment_Optimized-14            156106              7673 ns/op           17765 B/op        252 allocs/op
PASS
ok      github.com/OffchainLabs/prysm/v6/beacon-chain/rpc/prysm/v1alpha1/validator      3.133s
```